+++ CHANGELOG.md
@@ -1,9 +1,17 @@
 ## v.1.0.0
 
 ### Initial Release
+
 - Bult-in state-handling widget
 - Form state-handling support
 - Pull to Refresh support
 - Pagination support
 - State changes logging
 - Cubit wrapper
+
+## v.1.1.0
+
+### Local cache support
+
+- Adding local cache support in `BaseObjectLogic` & `BaseListLogic`
+- Fixing example code